Header -> HeaderAppBar; adjust header sizes; adjust margins in Home
